Manage concept structure prompt in PipeLLM, not in Prompt blueprint
